requirements.txt
pygame==2.1.2
azure-cognitiveservices-speech==1.25.0
openai==1.3.0
azure-core==1.26.4
python-dotenv==1.0.0